Split date type from date parsing
