+++ core/src/types/mod.rs
@@ -1,5 +1,7 @@
 mod bigint;
 mod bigrat;
+mod date;
 
 pub use bigint::{BigInt, BigIntError};
 pub use bigrat::BigRat;
+pub use date::GenericDateTime;
